+++ src/backend/opcode_parser.h
@@ -20,7 +20,7 @@ namespace athena::backend {
      * @return Position of next start value
      */
     unsigned int parse (
-            std::vector< vm_word > &bytecode, unsigned int &next_word,
+            std::vector< vm_word > &bytecode, unsigned long &next_word,
             OpCode &new_word, vm_word* args, unsigned int &argc
     );
 }
